Make microsoft accounts fully work
