+++ protocol/src/protocol/login.rs
@@ -1,4 +1,6 @@
-use super::{mojang::MojangAccount, UUID};
+use crate::protocol::microsoft::MicrosoftAccount;
+
+use super::UUID;
 use super::offline_acc::OfflineAccount;
 use dashmap::DashMap;
 use lazy_static::lazy_static;
@@ -23,14 +25,14 @@ pub trait AccountImpl {
 #[derive(Serialize, Deserialize)]
 pub struct Account {
     pub name: String,
-    pub uuid: Option<UUID>,
+    pub uuid: Option<String>,
     pub verification_tokens: Vec<String>, // this represents the verification tokens used to verify the account, such as hashed passwords, actual tokens, etc
     pub head_img_data: Option<Vec<u8>>,
     pub account_type: AccountType,
 }
 
 impl Account {
-    pub fn new(name: String, uuid: Option<UUID>, account_type: AccountType) -> Self {
+    pub fn new(name: String, uuid: Option<String>, account_type: AccountType) -> Self {
         Account {
             name,
             uuid,
@@ -111,7 +113,7 @@ lazy_static! {
         Arc::new({
             let map: DashMap<AccountType, Arc<dyn AccountImpl + Send + Sync>> = DashMap::new();
             // FIXME: These shouldn't be active all the time - someone might wanna disable them in a config file or something
-            map.insert(AccountType::Mojang, Arc::new(MojangAccount {}));
+            map.insert(AccountType::Microsoft, Arc::new(MicrosoftAccount {}));
             map.insert(AccountType::None, Arc::new(OfflineAccount {}));
             map
         });
@@ -119,7 +121,6 @@ lazy_static! {
 
 #[derive(Serialize, Deserialize, PartialEq, Eq, Hash, Clone)]
 pub enum AccountType {
-    Mojang, // FIXME: this has been deprecated, remove it
     Microsoft,
     Custom(String), // Not implemented yet, this will enable us to support other auth services without implementing every single one specifically
     None,           // aka. unverified or "offline account" (for offline mode servers)
